Лучшие изменения

+++ src/ru/elcus/mil/Mil1553RawPacketMT.java
@@ -113,9 +113,9 @@ package ru.elcus.mil;
  */
 public class Mil1553RawPacketMT {
 	public short[] basedata = new short[64];
-	public short sw;//mtgetsw();
-	public short statusword;//eventData.union.mt.wResultX;
-	public Mil1553RawPacketMT (short[] basedata,short sw,short statusword)
+	public int sw;//mtgetsw();
+	public int statusword;//eventData.union.mt.wResultX;
+	public Mil1553RawPacketMT (short[] basedata,int sw,int statusword)
 	{
 		this.sw = sw;
 		this.statusword = statusword;
